Add Filters.js file
All the exisitng filtering contents from script. js is moved to filters.js

in the future modifying filters could be done in this file, rather than the script.js file
