+++ index.html
@@ -101,8 +101,9 @@
             
         </div>
             <p id="donate_box"><span id ="donate"> <a href="https://github.com/webVerts#donate" target="_blank"><i>Donate Now</i></a></span> </p>
-
-        <script  src="resources/script.js"></script> 
+			
+        <script  src="resources/script.js" type="module"></script> 
+		
     </body>
 
 </html>
